refactor signer
