+++ crates/l2/contracts/Cargo.toml
@@ -27,6 +27,7 @@ ethrex-l2 = { path = "../../l2" }
 ethrex-sdk = { path = "../../l2/sdk" }
 ethrex-common = { path = "../../common" }
 ethrex-l2-common = { path = "../common" }
+ethrex-l2-rpc.workspace = true
 ethrex-rpc = { path = "../../networking/rpc" }
 genesis-tool = { path = "../../../tooling/genesis" }
