+++ crates/l2/networking/rpc/lib.rs
@@ -1,5 +1,7 @@
 pub mod l2;
 mod rpc;
 pub mod utils;
+pub mod signer;
+pub mod clients;
 
 pub use rpc::start_api;
