+++ crates/l2/sdk/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 ethrex-common.workspace = true
 ethrex-rpc.workspace = true
 ethrex-l2-common.workspace = true
+ethrex-l2-rpc.workspace = true
 
 ethereum-types.workspace = true
 tokio.workspace = true
