+++ crates/l2/sequencer/setup.rs
@@ -1,9 +1,10 @@
 use crate::sequencer::errors::ProofCoordinatorError;
-use ethrex_common::types::signer::{LocalSigner, Signer};
+use ethrex_l2_rpc::signer::{LocalSigner, Signer};
 use ethrex_common::{Address, Bytes};
 use ethrex_l2_common::calldata::Value;
 use ethrex_l2_sdk::calldata::encode_calldata;
 use ethrex_l2_sdk::get_address_from_secret_key;
+use ethrex_l2_rpc::clients::send_tx_bump_gas_exponential_backoff;
 use ethrex_rpc::clients::{
     Overrides,
     eth::{EthClient, WrappedTransaction},
@@ -88,8 +89,7 @@ pub async fn register_tdx_key(
 
     let signer = Signer::Local(LocalSigner::new(*private_key));
 
-    let verify_tx_hash = eth_client
-        .send_tx_bump_gas_exponential_backoff(&mut tx, &signer)
+    let verify_tx_hash = send_tx_bump_gas_exponential_backoff(eth_client, &mut tx, &signer)
         .await?;
 
     info!("Registered TDX key with transaction hash {verify_tx_hash:#x}");
