+++ crates/networking/rpc/clients/eth/mod.rs
@@ -22,8 +22,8 @@ use ethrex_common::{
     Address, H160, H256, Signature, U256,
     types::{
         BlobsBundle, Block, BlockHash, EIP1559Transaction, EIP4844Transaction, GenericTransaction,
-        PrivilegedL2Transaction, Signable, TxKind, TxType, WrappedEIP4844Transaction, batch::Batch,
-        block_execution_witness::ExecutionWitnessResult, signer::Signer,
+        PrivilegedL2Transaction, TxKind, TxType, WrappedEIP4844Transaction, batch::Batch,
+        block_execution_witness::ExecutionWitnessResult
     },
     utils::decode_hex,
 };
@@ -37,7 +37,6 @@ use secp256k1::SecretKey;
 use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
 use std::{ops::Div, str::FromStr};
-use tracing::warn;
 
 pub mod errors;
 pub mod eth_sender;
@@ -135,8 +134,6 @@ pub const BACKOFF_FACTOR: u64 = 2;
 // Give at least 8 blocks before trying to bump gas.
 pub const MIN_RETRY_DELAY: u64 = 96;
 pub const MAX_RETRY_DELAY: u64 = 1800;
-
-const WAIT_TIME_FOR_RECEIPT_SECONDS: u64 = 2;
 
 // 0x08c379a0 == Error(String)
 pub const ERROR_FUNCTION_SELECTOR: [u8; 4] = [0x08, 0xc3, 0x79, 0xa0];
@@ -282,157 +279,10 @@ impl EthClient {
         }
     }
 
-    pub async fn send_eip1559_transaction(
-        &self,
-        tx: &EIP1559Transaction,
-        signer: &Signer,
-    ) -> Result<H256, EthClientError> {
-        let signed_tx = tx.sign(signer).await?;
-
-        let mut encoded_tx = signed_tx.encode_to_vec();
-        encoded_tx.insert(0, TxType::EIP1559.into());
-
-        self.send_raw_transaction(encoded_tx.as_slice()).await
-    }
-
-    pub async fn send_eip4844_transaction(
-        &self,
-        wrapped_tx: &WrappedEIP4844Transaction,
-        signer: &Signer,
-    ) -> Result<H256, EthClientError> {
-        let mut wrapped_tx = wrapped_tx.clone();
-        wrapped_tx.tx.sign_inplace(signer).await?;
-
-        let mut encoded_tx = wrapped_tx.encode_to_vec();
-        encoded_tx.insert(0, TxType::EIP4844.into());
-
-        self.send_raw_transaction(encoded_tx.as_slice()).await
-    }
-
-    pub async fn send_wrapped_transaction(
-        &self,
-        wrapped_tx: &WrappedTransaction,
-        signer: &Signer,
-    ) -> Result<H256, EthClientError> {
-        match wrapped_tx {
-            WrappedTransaction::EIP4844(wrapped_eip4844_transaction) => {
-                self.send_eip4844_transaction(wrapped_eip4844_transaction, signer)
-                    .await
-            }
-            WrappedTransaction::EIP1559(eip1559_transaction) => {
-                self.send_eip1559_transaction(eip1559_transaction, signer)
-                    .await
-            }
-            WrappedTransaction::L2(privileged_l2_transaction) => {
-                self.send_privileged_l2_transaction(privileged_l2_transaction)
-                    .await
-            }
-        }
-    }
-
     /// Increase max fee per gas by percentage% (set it to (100+percentage)% of the original)
     pub fn bump_eip1559(&self, tx: &mut EIP1559Transaction, percentage: u64) {
         tx.max_fee_per_gas = (tx.max_fee_per_gas * (100 + percentage)) / 100;
         tx.max_priority_fee_per_gas += (tx.max_priority_fee_per_gas * (100 + percentage)) / 100;
-    }
-
-    pub async fn send_tx_bump_gas_exponential_backoff(
-        &self,
-        wrapped_tx: &mut WrappedTransaction,
-        signer: &Signer,
-    ) -> Result<H256, EthClientError> {
-        let mut number_of_retries = 0;
-
-        'outer: while number_of_retries < self.max_number_of_retries {
-            if let Some(max_fee_per_gas) = self.maximum_allowed_max_fee_per_gas {
-                let (tx_max_fee, tx_max_priority_fee) = match wrapped_tx {
-                    WrappedTransaction::EIP4844(tx) => (
-                        &mut tx.tx.max_fee_per_gas,
-                        &mut tx.tx.max_priority_fee_per_gas,
-                    ),
-                    WrappedTransaction::EIP1559(tx) => {
-                        (&mut tx.max_fee_per_gas, &mut tx.max_priority_fee_per_gas)
-                    }
-                    WrappedTransaction::L2(tx) => {
-                        (&mut tx.max_fee_per_gas, &mut tx.max_priority_fee_per_gas)
-                    }
-                };
-
-                if *tx_max_fee > max_fee_per_gas {
-                    *tx_max_fee = max_fee_per_gas;
-
-                    // Ensure that max_priority_fee_per_gas does not exceed max_fee_per_gas
-                    if *tx_max_priority_fee > *tx_max_fee {
-                        *tx_max_priority_fee = *tx_max_fee;
-                    }
-
-                    warn!(
-                        "max_fee_per_gas exceeds the allowed limit, adjusting it to {max_fee_per_gas}"
-                    );
-                }
-            }
-
-            // Check blob gas fees only for EIP4844 transactions
-            if let WrappedTransaction::EIP4844(tx) = wrapped_tx {
-                if let Some(max_fee_per_blob_gas) = self.maximum_allowed_max_fee_per_blob_gas {
-                    if tx.tx.max_fee_per_blob_gas > U256::from(max_fee_per_blob_gas) {
-                        tx.tx.max_fee_per_blob_gas = U256::from(max_fee_per_blob_gas);
-                        warn!(
-                            "max_fee_per_blob_gas exceeds the allowed limit, adjusting it to {max_fee_per_blob_gas}"
-                        );
-                    }
-                }
-            }
-            let tx_hash = self.send_wrapped_transaction(wrapped_tx, signer).await?;
-
-            if number_of_retries > 0 {
-                warn!(
-                    "Resending Transaction after bumping gas, attempts [{number_of_retries}/{}]\nTxHash: {tx_hash:#x}",
-                    self.max_number_of_retries
-                );
-            }
-
-            let mut receipt = self.get_transaction_receipt(tx_hash).await?;
-
-            let mut attempt = 1;
-            let attempts_to_wait_in_seconds = self
-                .backoff_factor
-                .pow(number_of_retries as u32)
-                .clamp(self.min_retry_delay, self.max_retry_delay);
-            while receipt.is_none() {
-                if attempt >= (attempts_to_wait_in_seconds / WAIT_TIME_FOR_RECEIPT_SECONDS) {
-                    // We waited long enough for the receipt but did not find it, bump gas
-                    // and go to the next one.
-                    match wrapped_tx {
-                        WrappedTransaction::EIP4844(wrapped_eip4844_transaction) => {
-                            self.bump_eip4844(wrapped_eip4844_transaction, 30);
-                        }
-                        WrappedTransaction::EIP1559(eip1559_transaction) => {
-                            self.bump_eip1559(eip1559_transaction, 30);
-                        }
-                        WrappedTransaction::L2(privileged_l2_transaction) => {
-                            self.bump_privileged_l2(privileged_l2_transaction, 30);
-                        }
-                    }
-
-                    number_of_retries += 1;
-                    continue 'outer;
-                }
-
-                attempt += 1;
-
-                tokio::time::sleep(std::time::Duration::from_secs(
-                    WAIT_TIME_FOR_RECEIPT_SECONDS,
-                ))
-                .await;
-
-                receipt = self.get_transaction_receipt(tx_hash).await?;
-            }
-
-            return Ok(tx_hash);
-        }
-
-        Err(EthClientError::TimeoutError)
     }
 
     /// Increase max fee per gas by percentage% (set it to (100+percentage)% of the original)
